# refactor: graph evaluation and dispatch composition