--- a/proto/internal/buf.yaml
+++ b/proto/internal/buf.yaml
@@ -1,0 +1,4 @@
+version: v1
+deps:
+  - buf.build/envoyproxy/protoc-gen-validate
+  - buf.build/openfga/api