… buf.gen.yaml
… go.sum
… internal/middleware/datastore/datastore.go
… pkg/cache/cache.go
… pkg/proto/dispatch/v1/dispatch.pb.go
… pkg/proto/dispatch/v1/dispatch.pb.validate.go
… pkg/proto/dispatch/v1/dispatch_grpc.pb.go
… proto/internal/buf.lock
… proto/internal/dispatch/v1/dispatch.proto
… server/server.go
… server/server_test.go
… server/test/check_query.go
… server/test/server.go