<!-- Thanks for opening a PR!  Here are some quick tips for you:
If this is your first time contributing, [read our Contributing Guidelines](https://github.com/openfga/openfga/blob/main/CONTRIBUTING.md) to learn how to create an acceptable PR for this repo.
By submitting a PR to this repository, you agree to the terms within the [OpenFGA Code of Conduct](https://github.com/openfga/openfga/blob/main/CODE_OF_CONDUCT.md)

If your PR is under active development, please submit it as a "draft". Once it's ready, open it up for review.
-->

<!-- Provide a brief summary of the changes -->

## Description
These changes introduce a new structure for resolving the relationship graph that is more flexible. These changes include a dispatch interface that allows us to inject behavior into the graph resolution process which will allow us to, for example, dispatch graph subproblems to other nodes in a cluster (if we so choose), integrate a dispatcher with a caching layer in it, and/or  delegate to a concurrent (but local) graph evaluation lookup (like we're simply doing today). We can also combine these dispatchers by chaining them together to implement a combined dispatcher that involves distributing requests to peers using something like consistent hashing in addition to caching.

I thought this implementation is much easier to understand and more elegant to compose than what we're doing today with the `resolutionContext`. Our `resolutionContext` strategy today is hard to understand because of the `rc.fork(..)` and shared state held across goroutines and synchronized by both channels and mutexes. This simplifies the implementation by better leveraging the composition of functions that take write channels, combining them with function chaining, and the use of `errgroup` to bring it all together.

## References
The implementation was largely inspired by the `graph` and `dispatch` packages in `authzed/spicedb`. 

## Review Checklist
- [ ] I have added documentation for new/changed functionality in this PR or in [openfga.dev](https://github.com/openfga/openfga.dev)
- [ ] The correct base branch is being used, if not `main`