@jon-whit 2022-06-14
[ ] - Wire up the resolution tracer so that we have the check resolution path.
[ ] - Add the metrics back into the graph resolution process.