@jon-whit 2022-07-19
Performed some preliminary benchmarks and found that these changes lead to additional heap allocations that caused a significant amount of GC pressure. Overall the performance of Check was diminished. Although the composition and readability improved, the performance degradation isn't worth the change.

For now I'm marking this as closed. It was a good exploration nonetheless.