@jon-whit 2022-07-20
For reference, the benchmarks that were run are defined below:

```
# after
➜  openfga git:(graph-dispatcher) ✗ go test -cpuprofile cpuprofile.out -memprofile memprofile.out -benchmem -count=5 -run=^$ -bench BenchmarkOpenFGAServer/TestMemoryDatastore/ ./server
goos: darwin
goarch: arm64
pkg: github.com/openfga/openfga/server
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   22642	     47808 ns/op	   40718 B/op	     641 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   32038	     37316 ns/op	   40575 B/op	     641 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   33388	     36142 ns/op	   40520 B/op	     641 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   35355	     35902 ns/op	   40673 B/op	     641 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   33944	     37152 ns/op	   40806 B/op	     641 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   24348	     45394 ns/op	   46992 B/op	     771 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   28252	     44576 ns/op	   45932 B/op	     771 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   20703	     54264 ns/op	   45917 B/op	     770 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   19468	     55117 ns/op	   45921 B/op	     770 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   19546	     55130 ns/op	   47994 B/op	     770 allocs/op

# before
➜  openfga git:(graph-dispatcher) ✗ go test -cpuprofile cpuprofile.out -memprofile memprofile.out -benchmem -count=5 -run=^$ -bench BenchmarkOpenFGAServer/TestMemoryDatastore/ ./server
goos: darwin
goarch: arm64
pkg: github.com/openfga/openfga/server
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   52742	     21984 ns/op	    9319 B/op	     186 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   56025	     21274 ns/op	    9332 B/op	     186 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   54915	     21379 ns/op	    9327 B/op	     186 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   56046	     20713 ns/op	    9329 B/op	     186 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkCheckWithoutTrace-10         	   57128	     20606 ns/op	    9333 B/op	     186 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   52317	     22205 ns/op	   11172 B/op	     261 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   53200	     22220 ns/op	   11185 B/op	     261 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   53659	     22236 ns/op	   11156 B/op	     260 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   53799	     21885 ns/op	   11167 B/op	     261 allocs/op
BenchmarkOpenFGAServer/TestMemoryDatastore/BenchmarkWithTrace-10                 	   53848	     22106 ns/op	   11171 B/op	     261 allocs/op
```

The number of heap allocations increased and the operations per second decreased with the changes introduced in this experimental PR. Using pprof to profile the differences, it appeared that the majority of the computation time was being bound to GC because of the heap allocations involved with the reduceable check functions. Although the code is easier to read, it probably doesn't warrant the refactor at this time because of this issue.